bone-age/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
joblib==1.2.0
numpy==1.21.6
pandas==1.3.5
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
typing-extensions==4.4.0
urllib3==1.26.13
